Add files via upload
To build a model to accurately detect the presence of Parkinson’s disease in an
individual.
Detecting Parkinson’s Disease with XGBoost – About the Python Machine
Learning Project
In this Python machine learning project, using the Python libraries scikit-learn,
numpy, pandas, and xgboost, we will build a model using an XGBClassifier. We’ll
load the data, get the features and labels, scale the features, then split the dataset,
build an XGBClassifier, and then calculate the accuracy of our model.